Fix RX handling when last bit was '1'
When last bit in the byte is '1' there are no more slopes in the byte so
next slope which can indicate that previous byte was sent is the first
slope of the next byte. It means that in such a situation the handler
must be invoked twice: firstly to finish the reception of the previous
byte and secondly to start reception of the new byte.