Test reception after going into idle state
When last bit of the byte was '1' there is no slope between the last
bit and the stop bit. If then there are no bits to be received the line
goes idle and the last byte won't be fully received until next slope.
This should be implemented in other way further. This test checks if the
bytes are received well in such a situation.